+++ CHANGES.md
@@ -49,6 +49,7 @@ _To Be Released..._
 
 **Fixes:**
 * [VIM-570](https://youtrack.jetbrains.com/issue/VIM-570) Print non-ascii characters in ex panel
+* [VIM-1958](https://youtrack.jetbrains.com/issue/VIM-1958) Fix `X` command for linewise selection
 
 0.55, 2020-01-20
 --------------